I fixed the problem by simply not loading the assets within the GameScreen class because I am only supposed to use its instance to access both the assetmanager and the spritebatch. Why that solved it? I am not sure, but it's best this way so that I can avoid making the assetmanager a singleton class which would act as if it was static and according to the assetmanager documentation, is undesirable.
I also added code to add a border around the pcg dungeon to enclose the world, but I decided to add it after the map iterations and this was because it could have messed up with the map iterations. Currently the dungeon world still needs to be worked on, especially in the case of disjointed rooms.